refactor: change app name to MainPage
