+++ src/components/MainPage.tsx
@@ -1,18 +1,18 @@
-import InputBox from './components/InputBox';
-import Task from './components/Task';
+import InputBox from './InputBox';
+import Task from './Task';
 import React, { useState } from 'react';
 import { DragDropContext, DropResult } from 'react-beautiful-dnd';
 
 export default function App() {
-  const [todo, setTodo] = useState <string> ("");
-  const [todos, setTodos] = useState <string[]> ([]);
-  const [completedTodos, setCompletedTodos] = useState <string[]> ([]);
+  const [todo, setTodo] = useState<string>('');
+  const [todos, setTodos] = useState<string[]>([]);
+  const [completedTodos, setCompletedTodos] = useState<string[]>([]);
 
   const handleAdd = (event: React.FormEvent) => {
     event.preventDefault();
-    if (todo !== "") {
+    if (todo !== '') {
       setTodos([...todos, todo]);
-      setTodo("");
+      setTodo('');
     }
   };
 
@@ -22,24 +22,27 @@ export default function App() {
     if (destination === null) {
       return;
     }
-    if (destination?.droppableId === source.droppableId && destination?.index === source.index) {
+    if (
+      destination?.droppableId === source.droppableId &&
+      destination?.index === source.index
+    ) {
       return;
     }
 
-    let tmp: string = '', currentTodos = [...todos], currentCompletedTodos = [...completedTodos];
+    let tmp: string = '',
+      currentTodos = [...todos],
+      currentCompletedTodos = [...completedTodos];
     if (source.droppableId === 'UnCompletedTodosList') {
       tmp = currentTodos[source.index];
       currentTodos.splice(source.index, 1);
-    }
-    else {
+    } else {
       tmp = currentCompletedTodos[source.index];
       currentCompletedTodos.splice(source.index, 1);
     }
 
     if (destination?.droppableId === 'UnCompletedTodosList') {
       currentTodos.splice(destination.index, 0, tmp);
-    }
-    else if (destination?.droppableId === 'CompletedTodosList') {
+    } else if (destination?.droppableId === 'CompletedTodosList') {
       currentCompletedTodos.splice(destination.index, 0, tmp);
     }
 
@@ -56,8 +59,13 @@ export default function App() {
         >
           Taskify V2
         </a>
-        <InputBox todo={todo} setTodo={setTodo} handleAdd={handleAdd}/>
-        <Task todos={todos} setTodos={setTodos} completedTodos={completedTodos} setCompletedTodos={setCompletedTodos}/> 
+        <InputBox todo={todo} setTodo={setTodo} handleAdd={handleAdd} />
+        <Task
+          todos={todos}
+          setTodos={setTodos}
+          completedTodos={completedTodos}
+          setCompletedTodos={setCompletedTodos}
+        />
       </main>
     </DragDropContext>
   );
